Implement flex to make divs into grid in container
Flex has been set to auto in the grid class and the children of the
container will are set to wrap. The width of a div is the
containers width divded by userInputs size pixels, with flex: auto
the div becomes a square and all the divs fit perfectly in the
container as a grid.