Make it so user can define grid size with button
Index.html was reworked a little to now be mostly in functions, besides
the initial querySelectors and event listeners. This was done to easily
call the each function as needed plus to cleanup a little.

The loop that created the grid was turned into the createGrid function.
That function has the same loop inside it, but take in size
as a parameter. The variables totalSquares and divWidth were moved to
this function and changed to use that size parameter instead of the
userInput varible, which has been removed. The function is called at the
end of the script to set the initial grid.

Function changeSize was added which will be called when the change
grid size button is clicked. It will ask the user for the size
they want and store it in inputSize, delete the current grid,
then call createGrid with inputSize as size.