Make it so divs randomly change color when hovered
The randomHex function is called in the background change for the div
that is currently hovered.
That function sets the max value of a hex color, then through
Math.random and multipling that by the max value + 1, the color value is
found. That value, currenly in base-10, is converted to a hexadecimal
string using toString; padded with 0's to have 6 characters; and then
made upper case which is then returned into that background change.