Have divs start at 10% opacity and work up to 100%
The darkness variable is assigned to 10 when page is loaded.
In the changeColor function an if statement was added to check if
darkness is less than 99. If it is then set the opacity of the event
target to darkness% and increase darkness by 10.
The changeSize function will reset darkness back to 10.